Noodling on ipsets
How much RAM do we need to store a bunch of unique IPv4 addresses
in a set in memory?  This commit contains a Python
script that reads a file containing one IPv4 address per line into
a set in memory and reports on the total RSS.  It's a simple quick
hack to give an idea of the overhead involved in basic data
structures with a large number of members.  To generate a file,
see bash oneliner in comments.