Add sample radix tree lookup implementation
This commit adds a sample radix tree test.  It reads the FireHOL
Level 1 ipset list into a radix tree using the py-radix module,
reports on the memory footprint consumed, the tests a few IP addresses
to see if they should be blocked or not (reporting a nanosecond time
for retrieval as it goes).  This is a basic proof of concept to see
how quick radix trees are for a real-world sample blocklist (the
Level 1 list is what FireHOL recommends most people use...it covers
around 620M addresses organized into 2751 subnets).